requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
pydantic==2.8.2
python-dateutil==2.9.0.post0
